+++ src/encode/pattern/mod.rs
@@ -60,10 +60,11 @@
 //! * `M`, `module` - The module that the log message came from, or `???` if not
 //!     provided.
 //! * `P`, `pid` - The current process id.
+//! * `i`, `tid` - The current system-wide unique thread ID.
 //! * `n` - A platform-specific newline.
 //! * `t`, `target` - The target of the log message.
 //! * `T`, `thread` - The name of the current thread.
-//! * `I`, `thread_id` - The ID of the current thread.
+//! * `I`, `thread_id` - The pthread ID of the current thread.
 //! * `X`, `mdc` - A value from the [MDC][MDC]. The first argument specifies
 //!     the key, and the second argument specifies the default value if the
 //!     key is not present in the MDC. The second argument is optional, and
@@ -149,10 +150,15 @@ fn default_color_map() -> HashMap<Level, Option<Color>> {
     color_map
 }
 
+thread_local!(
+    /// Thread-locally cached thread ID.
+    static TID: usize = thread_id::get()
+);
+
 /// The pattern encoder's configuration.
 #[cfg(feature = "config_parsing")]
-#[serde(deny_unknown_fields)]
 #[derive(Clone, Eq, PartialEq, Debug, Default, serde::Deserialize)]
+#[serde(deny_unknown_fields)]
 pub struct PatternEncoderConfig {
     #[serde(default = "default_pattern")]
     pattern: Option<String>,
@@ -429,10 +435,10 @@ impl<'a> From<Piece<'a>> for Chunk {
                     let timezone = match formatter.args.get(1) {
                         Some(arg) => {
                             if let Some(arg) = arg.get(0) {
-                                match arg {
-                                    Piece::Text(ref z) if *z == "utc" => Timezone::Utc,
-                                    Piece::Text(ref z) if *z == "local" => Timezone::Local,
-                                    Piece::Text(ref z) => {
+                                match *arg {
+                                    Piece::Text(z) if z == "utc" => Timezone::Utc,
+                                    Piece::Text(z) if z == "local" => Timezone::Local,
+                                    Piece::Text(z) => {
                                         return Chunk::Error(format!("invalid timezone `{}`", z));
                                     }
                                     _ => return Chunk::Error("invalid timezone".to_owned()),
@@ -475,6 +481,7 @@ impl<'a> From<Piece<'a>> for Chunk {
                 "T" | "thread" => no_args(&formatter.args, parameters, FormattedChunk::Thread),
                 "I" | "thread_id" => no_args(&formatter.args, parameters, FormattedChunk::ThreadId),
                 "P" | "pid" => no_args(&formatter.args, parameters, FormattedChunk::ProcessId),
+                "i" | "tid" => no_args(&formatter.args, parameters, FormattedChunk::SystemThreadId),
                 "t" | "target" => no_args(&formatter.args, parameters, FormattedChunk::Target),
                 "X" | "mdc" => {
                     if formatter.args.len() > 2 {
@@ -565,6 +572,7 @@ enum FormattedChunk {
     Thread,
     ThreadId,
     ProcessId,
+    SystemThreadId,
     Target,
     Newline,
     Align(Vec<Chunk>),
@@ -597,6 +605,9 @@ impl FormattedChunk {
             }
             FormattedChunk::ThreadId => w.write_all(thread_id::get().to_string().as_bytes()),
             FormattedChunk::ProcessId => w.write_all(process::id().to_string().as_bytes()),
+            FormattedChunk::SystemThreadId => {
+                TID.with(|tid| w.write_all(tid.to_string().as_bytes()))
+            }
             FormattedChunk::Target => w.write_all(record.target().as_bytes()),
             FormattedChunk::Newline => w.write_all(NEWLINE.as_bytes()),
             FormattedChunk::Align(ref chunks) => {
@@ -858,6 +869,18 @@ mod tests {
             .unwrap();
 
         assert_eq!(buf, process::id().to_string().as_bytes());
+    }
+
+    #[test]
+    #[cfg(feature = "simple_writer")]
+    fn system_thread_id() {
+        let pw = PatternEncoder::new("{i}");
+        let mut buf = vec![];
+
+        pw.encode(&mut SimpleWriter(&mut buf), &Record::builder().build())
+            .unwrap();
+
+        assert_eq!(buf, thread_id::get().to_string().as_bytes());
     }
 
     #[test]